flask/requirements.txt
Flask==2.2.3
PyYAML==6.0
Werkzeug==2.2.3

